修改了工程cmake文件,以使得工程可在windows下编译;修改了部分仅在linux下支持的函数,现已支持跨平台;修复了redis发布%b数据失败的问题,(由于设置的字符长度变量为int导致,改为size_t);修复了由于client_source的定时器没有完全初始化导致的windows下程序崩溃问题(因为windows不会默认初始化为0);删除了redis心跳的获取ip选项(该选项不够完善,可用配置文件替代)

+++ app/Koro_Caster_Service/include/Carrier/server_ntrip.h
@@ -34,7 +34,7 @@ private:
     bool _NtripVersion2 = false;
     bool _transfer_with_chunked = false;
 
-    int _chuncked_size = 0;
+    size_t _chuncked_size = 0;
 
     bufferevent *_bev;
     evbuffer *_evbuf;
