+++ app/Koro_Caster_Service/src/Carrier/client_ntrip.cpp
@@ -134,7 +134,7 @@ int client_ntrip::data_transfer(evbuffer *evbuf)
 
 int client_ntrip::publish_data_from_evbuf()
 {
-    int length = evbuffer_get_length(_evbuf);
+    size_t length = evbuffer_get_length(_evbuf);
     char *data = new char[length + 1];
     data[length] = '\0';
     evbuffer_remove(_evbuf, data, length);
